% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_api_state.R
\name{api_state}
\alias{api_state}
\title{Return the state of he Telraam API. Determine if updates can be made.}
\usage{
api_state(key)
}
\arguments{
\item{key}{file containing an API key ("key.txt")}
}
\value{
TRUE if the API responds well, FALSE otherwise
}
\description{
Return the state of he Telraam API. Determine if updates can be made.
}
