% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_decompose_2_fct.R
\name{decompose_2_fct}
\alias{decompose_2_fct}
\title{Decompose data from two sensors}
\usage{
decompose_2_fct(
  data,
  sensor1,
  sensor2,
  hour_x,
  direction1,
  direction2,
  mobility,
  norm,
  vacations = NULL,
  public_holidays = NULL,
  sensor_names = c("Burel-01", "Leclerc-02", "ParisMarche-03", "rueVignes-04",
    "ParisArcEnCiel-05", "RteVitre-06", "RueGdDomaine-07", "StDidierNord-08",
    "rueVallee-09", "StDidierSud-10", "RuePrieure-11", "RueCottage-12",
    "RueVeronniere-13", "RueDesEcoles-14"),
  sensor_ids = c(9000002156, 9000001906, 9000001618, 9000003090, 9000002453, 9000001844,
    9000001877, 9000002666, 9000002181, 9000002707, 9000003703, 9000003746, 9000003775,
    9000003736)
)
}
\arguments{
\item{data}{Data frame containing the data}

\item{sensor1}{The segment ID for sensor 1}

\item{sensor2}{The segment ID for sensor 2}

\item{hour_x}{The hour to filter on}

\item{direction1}{The direction for sensor 1}

\item{direction2}{The direction for sensor 2}

\item{mobility}{A character vector indicating the mobility types to include in the filter}

\item{norm}{A character indicating whether to normalize the results}

\item{vacations}{vacations period for filtering function}

\item{public_holidays}{public holidays period for filtering function}

\item{sensor_names}{A character vector containing the name of each sensor that is displayed to the user}

\item{sensor_ids}{A character vector containing the identifier name for each vector}
}
\value{
A list containing the decomposed data, seasonal data, peaks, and correlation
}
\description{
Decompose data from two sensors
}
