% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{filter_date}
\alias{filter_date}
\title{Select data within a specified date range}
\usage{
filter_date(data, date_range)
}
\arguments{
\item{data}{Data frame containing the date column}

\item{date_range}{A vector of two dates specifying the range}
}
\value{
Subset of the input data frame within the specified date range
}
\description{
Select data within a specified date range
}
\keyword{internal}
