% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{filter_public_holidays}
\alias{filter_public_holidays}
\title{Select data based on public holiday criteria}
\usage{
filter_public_holidays(data, JF)
}
\arguments{
\item{data}{Data frame containing the date column}

\item{JF}{A character indicating the public holiday selection criteria}
}
\value{
Subset of the input data frame based on the public holiday criteria
}
\description{
Select data based on public holiday criteria
}
\keyword{internal}
