% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{filter_vacation}
\alias{filter_vacation}
\title{Select data based on vacation criteria}
\usage{
filter_vacation(data, vacation)
}
\arguments{
\item{data}{Data frame containing the date column}

\item{vacation}{A character indicating the vacation selection criteria}
}
\value{
Subset of the input data frame based on the vacation criteria
}
\description{
Select data based on vacation criteria
}
\keyword{internal}
