% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{fine_tune_impute_missing_data}
\alias{fine_tune_impute_missing_data}
\title{Fine-tune Random Forest parameters for traffic data imputation}
\usage{
fine_tune_impute_missing_data(
  data,
  target_col = "vehicle",
  sensors_id = NULL,
  mtry_range,
  min_n_range,
  num_trees = 500,
  add_vars = NULL,
  threshold_uptime = 0.5
)
}
\arguments{
\item{data}{A data frame containing the traffic data.}

\item{target_col}{The name of the target column to impute. Default is "vehicle".}

\item{sensors_id}{An optional vector of sensor IDs to filter the data. Default is NULL (all sensors).}

\item{mtry_range}{A vector of mtry values to test.}

\item{min_n_range}{A vector of min.node.size values to test.}

\item{num_trees}{Integer. Number of trees to grow in the forest (default: 500).}

\item{add_vars}{An optional vector of additional variables to include in the model. Default is NULL.}

\item{threshold_uptime}{Numeric. The minimum uptime threshold for data inclusion (default: 0.5).}
}
\value{
A list containing the best parameters and metrics.
}
\description{
This function fine-tunes Random Forest parameters using Out-Of-Bag (OOB) validation to find the best hyperparameters for imputing missing traffic data.
}
\examples{
\dontrun{
fine_tune_impute_missing_data(data = traffic, target_col = "vehicle",
                        mtry_range = c(2, 4, 6, 8),
                        min_n_range = c(1, 5, 10, 20),
                        threshold_uptime = 0.5)
}
}
