% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{get_segment_name}
\alias{get_segment_name}
\title{Get the name of a segment giving its id}
\usage{
get_segment_name(segment_id)
}
\arguments{
\item{segment_id}{ID of segment, should be present in inst/config.yml}
}
\value{
Name of the segment, as specified in the configuration file, NULL otherwise.
}
\description{
Get the name of a segment giving its id
}
\keyword{internal}
