% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_gg_na_dist.R
\name{gg_na_dist}
\alias{gg_na_dist}
\title{gg_na_dist
The function gives a graphical representation of NA's distribution (monthly) during a selected period for one sensor.}
\usage{
gg_na_dist(
  data,
  sensor,
  start = NULL,
  end = NULL,
  hours = "ALL",
  list_weekday = "ALL",
  pub_holidays = "YES",
  holidays = "YES"
)
}
\arguments{
\item{data}{Data.Frame. A data frame containing road traffic data.}

\item{sensor}{Numeric. The sensor ID on which you want to see the representation.}

\item{start}{Character. The start of the period in the format "yyyy-mm-dd".}

\item{end}{Character. The end of the period in the format "yyyy-mm-dd".}

\item{hours}{Vector. A vector containing hours on which you want to see the representation, e.g., c("05h", "18h").}

\item{list_weekday}{Vector. A vector containing weekdays in French, e.g., c("Monday", "Sunday").}

\item{pub_holidays}{Character. "YES" if you want to see results with public holidays included;
"NO" if you want to exclude them from the representation;
"ONLY" if you only want to see them in the result.}

\item{holidays}{Character. "YES" if you want to see results including holidays;
"NO" if you want to exclude them from the representation;
"ONLY" if you only want to see them in the result.}
}
\value{
A ggplot graph.
}
\description{
gg_na_dist
The function gives a graphical representation of NA's distribution (monthly) during a selected period for one sensor.
}
