% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_gg_working_sensor.R
\name{gg_working_sensor}
\alias{gg_working_sensor}
\title{gg_working_sensor
the function gives a graphical representation of the sensors operation during a selected period,
the representation can be grouped or individual}
\usage{
gg_working_sensor(
  list_sensor = "ALL",
  data,
  start,
  end,
  hours = "ALL",
  aggregate = TRUE,
  list_weekday = "ALL",
  pub_holyday = "YES",
  holydays = "YES"
)
}
\arguments{
\item{list_sensor}{a vector containing sensor IDs, ex : c("9000001906", "9000003090)}

\item{data}{a data frame containing road traffic data}

\item{start}{the start of the period in the format "yyyy-mm-dd"}

\item{end}{the start of the period in the format "yyyy-mm-dd"}

\item{hours}{a vector containing hours on which you want to see the representation, ex : c("05h", "18h")}

\item{aggregate}{is aggregate = TRUE, you'll see information for all sensors together. if aggregate = FALSe,
the information are show sensor by sensor}

\item{list_weekday}{a vector containing weekdays in french, ex : c("lundi", "jeudi")}

\item{pub_holyday}{"YES" : if you want to see results with public holydays inclue;
"NO" if you want to exclude them from the representation;
"ONLY" if you only want to see them in the result}

\item{holydays}{"YES" : if you want to see results including holydays;
"NO" if you want to exclude them from the representation;
"ONLY" if you only want to see them in the result}
}
\value{
a ggplot graph
}
\description{
gg_working_sensor
the function gives a graphical representation of the sensors operation during a selected period,
the representation can be grouped or individual
}
