% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_import_sensor_comp.R
\name{import_sensor_comp}
\alias{import_sensor_comp}
\title{Import Sensor Data and Combine}
\usage{
import_sensor_comp(sensor_ids, sensor_comp_names)
}
\arguments{
\item{sensor_ids}{A character vector containing the sensor IDs to import.}

\item{sensor_comp_names}{A character vector containing the names of CSV files to import.}
}
\value{
A data frame containing the imported and combined sensor data.
}
\description{
This function imports sensor data from CSV files for given sensor IDs and combines them.
The function also processes the data by converting certain columns to lists.
}
