% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_na.R
\name{mutate_date_time}
\alias{mutate_date_time}
\title{Mutate Data}
\usage{
mutate_date_time(data, date, days, hours)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{date}{column date of data}

\item{days}{name of new column}

\item{hours}{name of new column}
}
\description{
This function summarizes the data by calculating the mean values of different variables
for each hour of the day.
}
