% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_working_sensor.R
\name{mutate_udh}
\alias{mutate_udh}
\title{Mutate Data}
\usage{
mutate_udh(data, uptime, date, upname, days, hours)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{uptime}{column uptime of data}

\item{date}{column date of data}

\item{upname}{name of the new column}

\item{days}{name of new column}

\item{hours}{name of new column}
}
\description{
This function summarizes the data by calculating the mean values of different variables
for each hour of the day.
}
