% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_comp_data.R
\name{mutate_uptime}
\alias{mutate_uptime}
\title{Mutate Data}
\usage{
mutate_uptime(data, uptime, upname)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{uptime}{column uptime of data}

\item{upname}{new column name}
}
\description{
This function summarizes the data by calculating the mean values of different variables
for each hour of the day.
}
