% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_na_prop.R
\name{na_prop}
\alias{na_prop}
\title{Visualisation NA stats}
\usage{
na_prop(data, start, end, sensor)
}
\arguments{
\item{data}{Data frame. Containing road traffic data.}

\item{start}{Character. The start date in the format "yyyy-mm-dd".}

\item{end}{Character. The end date in the format "yyyy-mm-dd".}

\item{sensor}{Character. The sensor ID on which you want to work.}
}
\value{
A character vector containing two phrases:
\enumerate{
\item The proportion of missing values before any data treatment.
\item The proportion of missing values replaced by imputed data and the remaining proportion of missing values.
}
}
\description{
This function calculates and returns the proportion of missing values (NA) and imputed values for a given sensor within a specified date range.
}
