% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_comparaison.R
\name{plot_comparaison}
\alias{plot_comparaison}
\title{Plot comparison between two datasets}
\usage{
plot_comparaison(param_ref, param_1, color_graph)
}
\arguments{
\item{param_ref}{Dataset of reference}

\item{param_1}{Dataset to compare}

\item{color_graph}{Color palette for the plot}
}
\value{
A ggplot object representing the comparison plot
}
\description{
Plot comparison between two datasets
}
