% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_plot_hour_threshold.R
\name{plot_hour_threshold}
\alias{plot_hour_threshold}
\title{Plot Eng React Function}
\usage{
plot_hour_threshold(...)
}
\arguments{
\item{...}{data and parameters to filter}
}
\value{
A list containing the plotly figure and the aggregated traffic data by hour.
}
\description{
This function plots the traffic data for a selected sensor, including the average speed and traffic volume
by hour of the day.
}
