% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plot_speed.R
\name{plot_speed}
\alias{plot_speed}
\title{Stock Function}
\usage{
plot_speed(...)
}
\arguments{
\item{...}{Arguments for filtering.}
}
\value{
A list containing the processed data and other results
}
\description{
Description of the function.
}
