% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plot_threshold.R
\name{plot_threshold}
\alias{plot_threshold}
\title{Plot Threshold on Speed Chart}
\usage{
plot_threshold(plot_speed, selected_speed, state_threshold, threshold)
}
\arguments{
\item{plot_speed}{A data frame containing the speed data and chart details.}

\item{selected_speed}{Character indicating the selected speed threshold.}

\item{state_threshold}{Character indicating the state of the threshold (automatique or manuel).}

\item{threshold}{Numeric value representing the manual threshold if state_threshold is "manuel".}
}
\value{
A ggplot2 chart with the speed plot and threshold line.
}
\description{
Plot Threshold on Speed Chart
}
