% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_comparaison.R
\name{prep}
\alias{prep}
\title{Prepare data for plotting}
\usage{
prep(data, count = "count", mean = "mean", var = "var", total = .data$total)
}
\arguments{
\item{data}{Data table to be processed}

\item{count}{Column name representing the count}

\item{mean}{Column name representing the mean}

\item{var}{Column name representing the variance}

\item{total}{Column total of table}
}
\value{
A processed data table for plotting
}
\description{
Prepare data for plotting
}
\keyword{internal}
