% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_comp_data.R
\name{prep_comp_data}
\alias{prep_comp_data}
\title{Prepares and completes road traffic data by replacing missing hour values with NA}
\usage{
prep_comp_data(data)
}
\arguments{
\item{data}{Data frame containing road traffic data.}
}
\value{
A complete data frame with all missing hour values replaced by NA.
}
\description{
This function prepares the input road traffic data by converting the date to a character format,
filtering out rows with missing dates, and performing additional data cleaning and imputation.
It also completes the data by adding missing hours for each sensor and converting the necessary
variables to numeric format.
}
