% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_gg_working_sensor.R
\name{prep_view_data}
\alias{prep_view_data}
\title{Prepare and filter data for viewing}
\usage{
prep_view_data(
  data,
  start = str_sub(min(data$date), 1, 10),
  end = str_sub(max(data$date), 1, 10),
  list_weekday = "ALL",
  pub_holidays = "YES",
  holidays = "YES"
)
}
\arguments{
\item{data}{The input data frame containing the sensor data.}

\item{start}{The start date in "yyyy-mm-dd" format (default is the minimum date in the data).}

\item{end}{The end date in "yyyy-mm-dd" format (default is the maximum date in the data).}

\item{list_weekday}{A character vector of weekdays to include in the analysis (default is "ALL").}

\item{pub_holidays}{A character indicating how to filter public holidays. Possible values are "YES", "NO", or "ONLY" (default is "YES").}

\item{holidays}{A character indicating how to filter holidays. Possible values are "YES", "NO", or "ONLY" (default is "YES").}
}
\value{
A data frame containing summarized information on operating sensors.
The data frame will have columns for sensor_id, days, and hourly uptime data (06h to 18h).
}
\description{
This function prepares and filters the input data based on specified parameters
}
