% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot_diagram_envelope.R, R/preprocessing.R
\name{replace_inactivity_period}
\alias{replace_inactivity_period}
\title{Replace or remove inactivity periods in traffic data}
\usage{
replace_inactivity_period(
  data,
  successive_day = 2,
  threshold_uptime = 0.5,
  remove_data = TRUE,
  show_graph = TRUE
)

replace_inactivity_period(
  data,
  successive_day = 2,
  threshold_uptime = 0.5,
  remove_data = TRUE,
  show_graph = TRUE
)
}
\arguments{
\item{data}{data.frame containing all the data for all your sensors}

\item{successive_day}{Integer. Number of consecutive days to define an inactivity period. Default is 2.}

\item{threshold_uptime}{Numeric. Uptime threshold chosen. Default is 0.5.}

\item{remove_data}{Logical. If TRUE, removes inactivity periods. If FALSE, replaces data with NA.}

\item{show_graph}{Logical. If TRUE, displays a graph of inactivity periods. Default is TRUE.}

\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{uptime_choice}{Real. Uptime choosen. Default to 0.5}

\item{retrieve_data}{Boolean. Choice to replace data or not. Default to TRUE}
}
\value{
enriched_data

data with inactivity periods replaced or removed
}
\description{
A short description...
Replace incomplete data with NA,

This function identifies periods of sensor inactivity and either replaces them with NA or removes them.
An inactivity period is defined as a consecutive period where the uptime is below the chosen threshold
or if no vehicles are counted during this period.
}
\examples{
replace_inactivity_period(traffic)
replace_inactivity_period(traffic,
 uptime_choice=0.3,
 successive_day=1,
 retrieve_data=FALSE)
traffic_without_inactivity <- replace_inactivity_period(traffic)
traffic_without_inactivity <- replace_inactivity_period(traffic,
                                                        threshold_uptime = 0.3,
                                                        successive_day = 1,
                                                        remove_data = FALSE,
                                                        show_graph = TRUE)
}
