% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot_diagram_envelope.R
\name{restore_v85}
\alias{restore_v85}
\title{Restore the v85 by direction from speed_hist_car}
\usage{
restore_v85(enriched_data, direction_choice)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}
}
\value{
enriched_data
}
\description{
Add two colums in the dataframe : v85_lft and v85_rgt
}
\examples{
restore_v85(enriched_data)
}
