% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot_diagram_envelope.R, R/preprocessing.R
\name{retrieve_missing_hours}
\alias{retrieve_missing_hours}
\title{Remove nighttime hours from traffic data}
\usage{
retrieve_missing_hours(data, threshold_uptime = 0.5)

retrieve_missing_hours(data, threshold_uptime = 0.5)
}
\arguments{
\item{data}{data.frame containing all the data for all your sensors}

\item{threshold_uptime}{Numeric. Uptime threshold chosen. Default is 0.5.}

\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{uptime_choice}{Real. Uptime choosen. Default to 0.5}
}
\value{
enriched_data

data without nighttime hours
}
\description{
A short description...
Retrieve hours with no data

This function removes nighttime hours when sensors cannot capture information.
It filters the data to retain only the hours with reliable uptime.
}
\examples{
retrieve_missing_hours(traffic)
retrieve_missing_hours(traffic,
 uptime_choice=0.3)
traffic_without_nighttime <- retrieve_missing_hours(traffic)
traffic_without_nighttime <- retrieve_missing_hours(traffic, threshold_uptime = 0.3)
}
