% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{set_global_vars}
\alias{set_global_vars}
\title{Function to set up the global variables for public holidays an vacations, with the default
being the french dates from a governmental API.}
\usage{
set_global_vars(vacations = NULL, public_holidays = NULL)
}
\arguments{
\item{vacations}{data frame containing the vacation dates}

\item{public_holidays}{data frame containing the public holidays dates}
}
\description{
Function to set up the global variables for public holidays an vacations, with the default
being the french dates from a governmental API.
}
\keyword{internal}
