% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{filtering}
\alias{filtering}
\title{Filter by selected criteria.
Not all criteria need to be filled in. Unfilled criteria are set by default so that no filtering is performed.}
\usage{
filtering(
  data = NULL,
  sensor = NULL,
  direction = " ",
  mobility = c("car", "heavy", "pedestrian", "bike"),
  date_range = NULL,
  vac = NULL,
  p_h = NULL,
  wkd = NULL,
  vacations = NULL,
  public_holidays = NULL
)
}
\arguments{
\item{data}{Raw data. See the "importation" function in the '2-import.R' file}

\item{sensor}{character. Name of desired sensor}

\item{direction}{character. Direction of the street: " " or "_lft" or "_rgt"}

\item{mobility}{character. Type of mobility: c("car","heavy","pedestrian","bike")}

\item{date_range}{character. Date or character. c("aaaa-mm-jj","aaaa-mm-jj")}

\item{vac}{character. With, without, or only with vacation: "Oui" or "Non" or "Seulement les vacances"}

\item{p_h}{character. With, without, or only with public holiday: "Oui" or "Non" or "Seulement les jours feries"}

\item{wkd}{character. Selected days of the week: c("1","2","3","4","5","6","7") here all days are selected}

\item{vacations}{vacation periods, set by default on the french ones}

\item{public_holidays}{public holidays period, set by default on the french ones}
}
\value{
the filtered data
}
\description{
Filter by selected criteria.
Not all criteria need to be filled in. Unfilled criteria are set by default so that no filtering is performed.
}
