% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_comparaison.R
\name{plot_comparaison}
\alias{plot_comparaison}
\title{Comparison time period}
\usage{
plot_comparaison(param_ref, param_1, color_graph)
}
\arguments{
\item{param_ref}{List of your data and parameters you want to filter on and compare, \link{filtering} : list(data=,sensor=,direction=,mobility=,date_range=,vac=,wkd=,p_h=)
Where: - data: dataframe containing your data
\itemize{
\item sensor: character, id of the sensor you want to study between quotes
\item direction: character, " " for all, "_rgt" for right (B to A), "_lft" for left (A to B)
\item mobility: character vector, c("car", "heavy", "pedestrian", "bike") are all the types you can choose from
\item date_range: date vector, c("starting_date", "ending_date")
\item vac: character, "YES" to keep, "NO" to filter out, or "ONLY" to only take vacation in the database
\item wkd: character vector, c("1", "2", "3", "4", "5", "6", "7"), where "1" stands for Monday and "7" for Sunday. Use options(lubridate.week.start = 1) for this function to perform correctly.
\item p_h: character, "YES" to keep, "NO" to filter out, or "ONLY" to only take public holidays in the database
}}

\item{param_1}{List, same as param_ref}

\item{color_graph}{Color palette for the plot}
}
\value{
A ggplot object representing the comparison plot
}
\description{
Comparative visualization of the number of users for each hour of the day there is data for over two different time periods.
}
